Add files via upload
 Simple AI-Powered Text Classifier 

This Python app uses TextBlob to classify the sentiment of a sentence as Positive, Negative, or Neutral based on polarity scores. 
The interface is built using Gradio, allowing users to input text directly in a simple web interface. 
Each classification is also saved to a local text file for record-keeping.